fix: ROOT CAUSE FOUND - Token IDs must be uint32, not float32
BREAKTHROUGH: Found and fixed the root cause of the batch processing bug.

ROOT CAUSE:
Token IDs were being passed as float32 instead of uint32/int32.
ttnn::embedding does NOT correctly handle float32 inputs for batched
processing - it returns identical outputs for all batch samples.

THE BUG WAS NOT IN BERT OR ttnn::embedding ITSELF:
- ttnn::embedding works correctly with proper uint32 inputs
- BERT implementation is correct
- The bug was in TEST CODE using wrong dtypes

EVIDENCE:
Tests with uint32 (correct):
- BertEmbeddingPipelineTest: PASS (-1.52 vs -0.94)
- BroadcastingHypothesisTest: PASS (4.47 vs 63.25)
- All manual pipeline tests: ALL PASS

Tests with float32 (WRONG):
- BertBatchIsolationTest (before fix): FAIL (identical: -1.52 vs -1.52)
- BertBatchBugTest (before fix): FAIL (identical outputs)
- Python debug script (before fix): FAIL (identical outputs)

Tests with uint32 (AFTER FIX):
- BertBatchIsolationTest: PASS (-1.52 vs -0.94)
- BertBatchBugTest: PASS (-0.0032 vs -0.0028)
- Python debug script: PASS (-0.195 vs -0.043)

CHANGES:
1. Removed workaround from embedding_op.cpp (restored original simple code)
2. Fixed BertBatchBugTest to use uint32_t for token IDs
3. Fixed BertBatchIsolationTest to use uint32_t for token IDs
4. Fixed debug_batch_processing.py to use np.uint32 for token IDs

VALIDATION:
All tests now pass WITHOUT any workarounds:
- C++ tests: ALL PASS with uint32_t
- Python tests: PASS with np.uint32
- Batch processing works correctly
- No performance overhead from workaround

Files changed:
- sources/ttml/ops/embedding_op.cpp: Removed workaround
- tests/model/bert_batch_bug_test.cpp: Fixed to use uint32_t
- tests/model/bert_batch_isolation_test.cpp: Fixed to use uint32_t
- debug_batch_processing.py: Fixed to use np.uint32

+++ tt-train/debug_batch_processing.py
@@ -148,8 +148,9 @@ def test_batch_independence():
     print(f"HF outputs different? {not np.allclose(hf_logits[0], hf_logits[1], rtol=1e-3)}")
 
     # TTML forward pass
-    input_ids_reshaped = input_ids_np.reshape(batch_size, 1, 1, seq_len).astype(np.float32)
-    token_type_ids_reshaped = token_type_ids_np.reshape(batch_size, 1, 1, seq_len).astype(np.float32)
+    # IMPORTANT: Token IDs must be integers (uint32/int32), NOT float32!
+    input_ids_reshaped = input_ids_np.reshape(batch_size, 1, 1, seq_len).astype(np.uint32)
+    token_type_ids_reshaped = token_type_ids_np.reshape(batch_size, 1, 1, seq_len).astype(np.uint32)
     attention_mask_reshaped = attention_mask_np.reshape(batch_size, 1, 1, seq_len).astype(np.float32)
 
     print(f"\nTTML input shape: {input_ids_reshaped.shape}")
